fix the logic

+++ src/services/apiCabins.js
@@ -1,53 +1,68 @@
-import supabase, { supabaseUrl } from './superbase';
+import supabase, { supabaseUrl } from "./supabase";
 
-export async function getCabin() {
-  const { data, error } = await supabase.from('cabins').select('*');
+export async function getCabins() {
+  const { data, error } = await supabase.from("cabins").select("*");
 
   if (error) {
-    console.error('cabins could not be laoded');
-    throw new Error('failed to laod the cabins data');
+    console.error(error);
+    throw new Error("Cabins could not be loaded");
   }
 
   return data;
 }
 
-export const createCabin = async function (newCabin) {
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
+
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
-    '/',
-    ''
+    "/",
+    ""
   );
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  // 1. Create/edit cabin
+  let query = supabase.from("cabins");
 
-  const { data, error } = await supabase
-    .from('cabins')
-    .insert([{ ...newCabin, image: imagePath }]);
+  // A) CREATE
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+  // B) EDIT
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+
+  const { data, error } = await query.select().single();
+
   if (error) {
-    console.error('cabins could not be created');
-    throw new Error('failed to create the cabins');
+    console.error(error);
+    throw new Error("Cabin could not be created");
   }
 
+  // 2. Upload image
+  if (hasImagePath) return data;
+
   const { error: storageError } = await supabase.storage
-    .from('cabin-images')
+    .from("cabin-images")
     .upload(imageName, newCabin.image);
 
+  // 3. Delete the cabin IF there was an error uplaoding image
   if (storageError) {
-    await supabase.from('cabins').delete().eq('id', data.id);
+    await supabase.from("cabins").delete().eq("id", data.id);
     console.error(storageError);
     throw new Error(
-      'cabin imagae could not be created hence cabin could not be uploaded  '
+      "Cabin image could not be uploaded and the cabin was not created"
     );
   }
 
   return data;
-};
+}
 
 export async function deleteCabin(id) {
-  const { data, error } = await supabase.from('cabins').delete().eq('id', id);
+  const { data, error } = await supabase.from("cabins").delete().eq("id", id);
 
   if (error) {
-    console.error('cabins could not be deleted');
-    throw new Error('failed to delete the cabins data');
+    console.error(error);
+    throw new Error("Cabin could not be deleted");
   }
 
   return data;
